+++ core/types/governance_update_spot_market.go
@@ -122,19 +122,16 @@ func (n UpdateSpotMarket) DeepClone() *UpdateSpotMarket {
 }
 
 type UpdateSpotMarketConfiguration struct {
-	Instrument                *UpdateInstrumentConfiguration
 	Metadata                  []string
 	PriceMonitoringParameters *PriceMonitoringParameters
 	TargetStakeParameters     *TargetStakeParameters
 	RiskParameters            updateRiskParams
 	SLAParams                 *LiquiditySLAParams
-	MarkPriceConfiguration    *CompositePriceConfiguration
 }
 
 func (n UpdateSpotMarketConfiguration) String() string {
 	return fmt.Sprintf(
-		"instrument(%s) metadata(%v) priceMonitoring(%s) targetStakeParameters(%s) risk(%s) slaParams(%s)",
-		stringer.PtrToString(n.Instrument),
+		"metadata(%v) priceMonitoring(%s) targetStakeParameters(%s) risk(%s) slaParams(%s)",
 		MetadataList(n.Metadata).String(),
 		stringer.PtrToString(n.PriceMonitoringParameters),
 		stringer.PtrToString(n.TargetStakeParameters),
@@ -145,14 +142,10 @@ func (n UpdateSpotMarketConfiguration) String() string {
 
 func (n UpdateSpotMarketConfiguration) DeepClone() *UpdateSpotMarketConfiguration {
 	cpy := &UpdateSpotMarketConfiguration{
-		Metadata:               make([]string, len(n.Metadata)),
-		SLAParams:              n.SLAParams.DeepClone(),
-		MarkPriceConfiguration: n.MarkPriceConfiguration.DeepClone(),
+		Metadata:  make([]string, len(n.Metadata)),
+		SLAParams: n.SLAParams.DeepClone(),
 	}
 	cpy.Metadata = append(cpy.Metadata, n.Metadata...)
-	if n.Instrument != nil {
-		cpy.Instrument = n.Instrument.DeepClone()
-	}
 	if n.PriceMonitoringParameters != nil {
 		cpy.PriceMonitoringParameters = n.PriceMonitoringParameters.DeepClone()
 	}
@@ -181,7 +174,6 @@ func (n UpdateSpotMarketConfiguration) IntoProto() *vegapb.UpdateSpotMarketConfi
 		PriceMonitoringParameters: priceMonitoring,
 		TargetStakeParameters:     targetStakeParameters,
 		SlaParams:                 n.SLAParams.IntoProto(),
-		MarkPriceConfiguration:    n.MarkPriceConfiguration.IntoProto(),
 	}
 	switch rp := riskParams.(type) {
 	case *vegapb.UpdateSpotMarketConfiguration_Simple:
@@ -211,7 +203,6 @@ func UpdateSpotMarketConfigurationFromProto(p *vegapb.UpdateSpotMarketConfigurat
 		PriceMonitoringParameters: priceMonitoring,
 		TargetStakeParameters:     targetStakeParameters,
 		SLAParams:                 slaParams,
-		MarkPriceConfiguration:    CompositePriceConfigurationFromProto(p.MarkPriceConfiguration),
 	}
 	if p.RiskParameters != nil {
 		switch rp := p.RiskParameters.(type) {