+++ CHANGELOG.md
@@ -233,6 +233,8 @@
 - [10604](https://github.com/vegaprotocol/vega/issues/10604) - Register margin modes API subscriber.
 - [10595](https://github.com/vegaprotocol/vega/issues/10595) - Fix failed amends for isolated margin orders causing negative spread in console.
 - [10606](https://github.com/vegaprotocol/vega/issues/10606) - Party profiles `API` was not returning results.
+- [10625](https://github.com/vegaprotocol/vega/issues/10625) - Fix panic in update spot market. 
+- [9244](https://github.com/vegaprotocol/vega/issues/9244) - fix unit tests to use spot markets. 
 - [10625](https://github.com/vegaprotocol/vega/issues/10625) - Fix panic in update spot market.
 - [10649](https://github.com/vegaprotocol/vega/issues/10649) - Ensure markets do not get stuck in liquidity auction after protocol upgrade.